% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{adf}
\alias{adf}
\title{Return the p-value of the Augmented Dickey-Fuller test.}
\usage{
adf(x)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}
}
\value{
A p-value.
}
\description{
Return the p-value of the Augmented Dickey-Fuller test.
}
\examples{
\dontrun{adf(x)}
}
