% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{diff_series}
\alias{diff_series}
\title{Applies a fractional differentiation to a time series.}
\usage{
diff_series(x, d)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}

\item{d}{Order of the fractional differentiation.}
}
\value{
A time series object.
}
\description{
Applies a fractional differentiation to a time series.
}
\examples{
\dontrun{diff_series(x, 0.5)}
}
