% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{grid_d}
\alias{grid_d}
\title{Returns a grid for different values of the fractional differentiation order parameter `d`.}
\usage{
grid_d(x, fun, dFrom = 0, dTo = 1, dBy = 0.1)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}

\item{fun}{A function taking as only argument the time series and returning a p-value.}

\item{dFrom}{Lower bound for `d`.}

\item{dTo}{Upper bound for `d`.}

\item{dBy}{Increment for each in the grid.}
}
\value{
A grid with four columns: the value of the fractional differentiation order `d`, the p-value returned by `fun`, the Pearson correlation coefficient between the original and the transform series, and the Spearman counterpart.
}
\description{
Returns a grid for different values of the fractional differentiation order parameter `d`.
}
\examples{
\dontrun{grid_d(x, adf)}
}
