% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{kpss}
\alias{kpss}
\title{Return the p-value of the Kwiatkowski–Phillips–Schmidt–Shin test.}
\usage{
kpss(x)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}
}
\value{
A p-value.
}
\description{
Return the p-value of the Kwiatkowski–Phillips–Schmidt–Shin test.
}
\examples{
\dontrun{kpss(x)}
}
