% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{min_d}
\alias{min_d}
\title{Finds the least fractional differentiation order subject to fun returning less than threshold.}
\usage{
min_d(x, fun, threshold, dFrom = 0, dTo = 1, dBy = 0.1)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}

\item{fun}{A function taking as only argument the time series and returning a p-value.}

\item{threshold}{A numerical threshold for the p-value.}

\item{dFrom}{Lower bound for `d`.}

\item{dTo}{Upper bound for `d`.}

\item{dBy}{Increment for each in the grid.}
}
\value{
The least value of `d`.
}
\description{
Finds the least fractional differentiation order subject to fun returning less than threshold.
}
\examples{
\dontrun{min_d(x, adf, 0.05)}
}
