% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{roll_min_d}
\alias{roll_min_d}
\title{Finds the least fractional differentiation order subject to fun returning less than threshold for a rolling window.}
\usage{
roll_min_d(x, width, by, ...)
}
\arguments{
\item{x}{A time series. Either a numberic vector or an `xts` object.}

\item{width}{The number of observations included in each window.}

\item{by}{The number of observations between each run.}

\item{...}{Other parameters to be passed to zoo::rollapply and min_d.}
}
\value{
An `xts` object with the time series of `d`.
}
\description{
Finds the least fractional differentiation order subject to fun returning less than threshold for a rolling window.
}
\examples{
\dontrun{roll_min_d(x, width = 252, by = 1)}
}
